Moved sound system initialization to Context, made it warn on failure.
Eventually I want to make all initialization optional, and have you
able to pass a list of required modules...
